Reworked logging and added support for manual acks.
The winston NPM module is now being used for logging.
The SDK's log level is now supported by the config.
When a push websocket is established, it can now be configured not to perform
automatic acknowledges. There is a new function on the push websocket to
acknowledge messages manually by ID.